+++ src/app/dashboard/bot-widget/bot-widget.component.ts
@@ -18,7 +18,7 @@ export class BotWidgetComponent implements OnInit {
     size: new FormControl('large')
   });
 
-  get widgetURL() { return `${environment.url}/api/bots/${this.id}/widget?size=${this.form.value.size}`; }
+  get widgetURL() { return `${environment.url}/api/v1/bots/${this.id}/widget?size=${this.form.value.size}`; }
   get id() { return this.route.snapshot.paramMap.get('id') }
 
   constructor(